backend/requirements.txt
asgiref==3.8.1
csrf==0.1b1
Django==5.0.7
django-cors-headers==4.4.0
django-decouple==2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.5.1
typing_extensions==4.12.2
